Extracted build_name stuff
